Added song remove + Song filename display
Song remove is fairly simple - Addition of new PLAYER_CTRL for removing
a song, and the position of the song in the playlist is passed.

Song filename display is a bit more complex. The solution is to add
another argument type specefically for 'song's. Each song has two
different cons_printf_compiled formats, and one or the other is used
depending on what name information is aviliable for that song. At the
moment it is good enough to display the standard triple, and then
fallback to the filename. Both the triple and filename display are
configurable separately from eachother.